requirements/base.txt
Django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg==3.2.1
celery==5.4.0
redis==5.0.8
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
paypalrestsdk==1.13.3
django-celery-beat==2.7.0
